+++ tests/phpunit/integration/api/actions/beansModifyAction.php
@@ -9,7 +9,9 @@
 
 namespace Beans\Framework\Tests\Integration\API\Actions;
 
-use WP_UnitTestCase;
+use Beans\Framework\Tests\Integration\API\Actions\Includes\Actions_Test_Case;
+
+require_once __DIR__ . '/includes/class-actions-test-case.php';
 
 /**
  * Class Tests_BeansModifyAction
@@ -18,7 +20,7 @@ use WP_UnitTestCase;
  * @group   unit-integration
  * @group   api
  */
-class Tests_BeansModifyAction extends WP_UnitTestCase {
+class Tests_BeansModifyAction extends Actions_Test_Case {
 
 	/**
 	 * Reset the test fixture.
@@ -116,68 +118,5 @@ class Tests_BeansModifyAction extends WP_UnitTestCase {
 		$this->assertEquals( $modified_action, _beans_get_action( 'beans', 'modified' ) );
 		$this->assertTrue( has_action( $action['hook'] ) );
 		$this->check_parameters_registered_in_wp( array_merge( $action, $modified_action ) );
-	}
-
-	/**
-	 * Check that it is not registered first.
-	 *
-	 * @since 1.5.0
-	 *
-	 * @param string $id   The ID to check.
-	 * @param string $hook The hook (event name) to check.
-	 *
-	 * @return void
-	 */
-	protected function check_not_added( $id, $hook ) {
-		$this->assertFalse( _beans_get_action( $id, 'added' ) );
-		$this->assertFalse( has_action( $hook ) );
-	}
-
-	/**
-	 * Check that the right parameters are registered in WordPress.
-	 *
-	 * @since 1.5.0
-	 *
-	 * @param array $action The action that should be registered.
-	 *
-	 * @return void
-	 */
-	protected function check_parameters_registered_in_wp( array $action ) {
-		global $wp_filter;
-		$registered_action = $wp_filter[ $action['hook'] ]->callbacks[ $action['priority'] ];
-
-		$this->assertArrayHasKey( $action['callback'], $registered_action );
-		$this->assertEquals( $action['callback'], $registered_action[ $action['callback'] ]['function'] );
-		$this->assertEquals( $action['args'], $registered_action[ $action['callback'] ]['accepted_args'] );
-
-		// Then remove the action.
-		remove_action( $action['hook'], $action['callback'], $action['priority'] );
-	}
-
-	/**
-	 * Setup the original action.
-	 *
-	 * @since 1.5.0
-	 *
-	 * @param string $id Optional. Beans ID to register. Default is 'foo'.
-	 *
-	 * @return array
-	 */
-	protected function setup_original_action( $id = 'foo' ) {
-		$action = array(
-			'hook'     => "{$id}_hook",
-			'callback' => "callback_{$id}",
-			'priority' => 10,
-			'args'     => 1,
-		);
-
-		$this->check_not_added( $id, $action['hook'] );
-
-		// Add the original action to get us rolling.
-		beans_add_action( $id, $action['hook'], $action['callback'] );
-		$this->assertTrue( has_action( $action['hook'] ) );
-		$this->check_parameters_registered_in_wp( $action );
-
-		return $action;
 	}
 }
