+++ tests/phpunit/unit/api/actions/includes/class-actions-test-case.php
@@ -1,0 +1,95 @@
+<?php
+/**
+ * Tests Case for Beans' Action API unit tests.
+ *
+ * @package Beans\Framework\Tests\Unit\API\Actions\Includes
+ *
+ * @since   1.5.0
+ */
+
+namespace Beans\Framework\Tests\Unit\API\Actions\Includes;
+
+use Beans\Framework\Tests\Unit\Test_Case;
+use Brain\Monkey;
+
+/**
+ * Abstract Class Actions_Test_Case
+ *
+ * @package Beans\Framework\Tests\Unit\API\Actions\Includes
+ */
+abstract class Actions_Test_Case extends Test_Case {
+
+	/**
+	 * Setup test fixture.
+	 */
+	protected function setUp() {
+		parent::setUp();
+
+		require_once BEANS_TESTS_LIB_DIR . 'api/actions/functions.php';
+		require_once BEANS_TESTS_LIB_DIR . 'api/utilities/functions.php';
+	}
+
+	/**
+	 * Reset the test fixture.
+	 */
+	protected function tearDown() {
+		parent::tearDown();
+
+		global $_beans_registered_actions;
+		$_beans_registered_actions = array(
+			'added'    => array(),
+			'modified' => array(),
+			'removed'  => array(),
+			'replaced' => array(),
+		);
+	}
+
+	/**
+	 * Check that is not registered first.
+	 *
+	 * @since 1.5.0
+	 *
+	 * @param string $id   The ID to check.
+	 * @param string $hook The hook (event name) to check.
+	 *
+	 * @return void
+	 */
+	protected function check_not_added( $id, $hook ) {
+		$this->assertFalse( _beans_get_action( $id, 'added' ) );
+		$this->assertFalse( has_action( $hook ) );
+	}
+
+	/**
+	 * Setup the original action.
+	 *
+	 * @since 1.5.0
+	 *
+	 * @param string $id Optional. Beans ID to register. Default is 'foo'.
+	 *
+	 * @return array
+	 */
+	protected function setup_original_action( $id = 'foo' ) {
+		$container = Monkey\Container::instance();
+		$action    = array(
+			'hook'     => "{$id}_hook",
+			'callback' => "callback_{$id}",
+			'priority' => 10,
+			'args'     => 1,
+		);
+
+		$this->check_not_added( $id, $action['hook'] );
+
+		// Add the original action to get us rolling.
+		beans_add_action( $id, $action['hook'], $action['callback'] );
+		$this->assertTrue( has_action( $action['hook'] ) );
+		$this->assertTrue(
+			$container->hookStorage()->isHookAdded(
+				Monkey\Hook\HookStorage::ACTIONS,
+				$action['hook'],
+				$action['callback']
+			)
+		);
+
+		return $action;
+	}
+}
